refactor ChapterOpeningOtherUser
